+++ sdk/servicebus/azure-messaging-servicebus-stress/src/main/java/com/azure/messaging/servicebus/stress/scenarios/MessageSessionProcessor.java
@@ -4,14 +4,14 @@
 package com.azure.messaging.servicebus.stress.scenarios;
 
 import com.azure.core.util.logging.ClientLogger;
-import com.azure.messaging.servicebus.ServiceBusClientBuilder;
 import com.azure.messaging.servicebus.ServiceBusProcessorClient;
-import com.azure.messaging.servicebus.models.ServiceBusReceiveMode;
-import com.azure.messaging.servicebus.stress.util.EntityType;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
-import java.util.concurrent.CountDownLatch;
+import java.time.Duration;
+
+import static com.azure.messaging.servicebus.stress.scenarios.TestUtils.blockingWait;
+import static com.azure.messaging.servicebus.stress.scenarios.TestUtils.getSessionProcessorBuilder;
 
 /**
  * Test ServiceBusSessionProcessorClient
@@ -19,6 +19,9 @@ import java.util.concurrent.CountDownLatch;
 @Component("MessageSessionProcessor")
 public class MessageSessionProcessor extends ServiceBusScenario {
     private static final ClientLogger LOGGER = new ClientLogger(MessageSessionProcessor.class);
+
+    @Value("${DURATION_IN_MINUTES:15}")
+    private int durationInMinutes;
 
     @Value("${MAX_CONCURRENT_SESSIONS:1}")
     private int maxConcurrentSessions;
@@ -31,57 +34,18 @@ public class MessageSessionProcessor extends ServiceBusScenario {
 
     @Override
     public void run() {
-        final CountDownLatch latch = new CountDownLatch(1);
-
-        final String connectionString = options.getServicebusConnectionString();
-        final EntityType entityType = options.getServicebusEntityType();
-        String queueName = null;
-        String topicName = null;
-        String subscriptionName = null;
-        if (entityType == EntityType.QUEUE) {
-            queueName = options.getServicebusSessionQueueName();
-        } else if (entityType == EntityType.TOPIC) {
-            topicName = options.getServicebusTopicName();
-            subscriptionName = options.getServicebusSessionSubscriptionName();
-        }
-
-        ServiceBusProcessorClient client = new ServiceBusClientBuilder()
-                .connectionString(connectionString)
-                .sessionProcessor()
-                .queueName(queueName)
-                .topicName(topicName)
-                .subscriptionName(subscriptionName)
+        ServiceBusProcessorClient processor = getSessionProcessorBuilder(options)
                 .maxConcurrentSessions(maxConcurrentSessions)
                 .maxConcurrentCalls(maxConcurrentCalls)
                 .prefetchCount(prefetchCount)
-                .receiveMode(ServiceBusReceiveMode.PEEK_LOCK)
-                .disableAutoComplete()
-                .processMessage(messageContext -> {
-                    LOGGER.verbose("Before complete. messageId: {}, sessionId: {}",
-                            messageContext.getMessage().getMessageId(),
-                            messageContext.getMessage().getSessionId());
-                    messageContext.complete();
-                    LOGGER.verbose("After complete. messageId: {}, sessionId: {}",
-                            messageContext.getMessage().getMessageId(),
-                            messageContext.getMessage().getSessionId());
-                })
+                .processMessage(messageContext -> messageContext.complete())
                 .processError(err -> {
                     throw LOGGER.logExceptionAsError(new RuntimeException(err.getException()));
                 })
                 .buildProcessorClient();
 
-        client.start();
-
-        // When the connection is recovering, there is a gap between the creation of new reactor-executor thread and
-        // the disposal of old reactor-executor thread. Since only the daemon threads are running, the program ends.
-        // Here we add a 'CountDownLatch' to block main thread and keep the processor running forever.
-        // In the future, we can add wait time as test parameter so that we can control the testing time.
-        try {
-            latch.await();
-        } catch (InterruptedException e) {
-            e.printStackTrace();
-        }
-        // We won't hit here unless we add a wait time for the 'CountDownLatch'.
-        client.close();
+        processor.start();
+        blockingWait(Duration.ofMinutes(durationInMinutes));
+        processor.close();
     }
 }